Various clean up
- Add has_one test cases
- update default order of has_one/belong_to relation to match the rest
- Removed unused factory bot
- Let test suites define their own schema/models for the test
