+++ test/sort/has_and_belongs_to_many_test.rb
@@ -1,47 +1,30 @@
 require 'test_helper'
 
-class SortRelationTest < ActiveSupport::TestCase
+class HasAndBelongsToManySortTest < ActiveSupport::TestCase
 
-  test '::sort(:has_many_relationship => :column)' do
-    query = Property.sort(:addresses => :id)
-
-    assert_equal(<<-SQL.strip.gsub(/\s+/, ' '), query.to_sql.gsub(/\s+/, ' '))
-      SELECT "properties".*, "addresses"."id" FROM "properties"
-      INNER JOIN "addresses" ON "addresses"."property_id" = "properties"."id"
-      ORDER BY "addresses"."id" ASC
-    SQL
+  schema do
+    create_table "properties", force: :cascade do |t|
+      t.string   "name",                 limit: 255
+    end
+    
+    create_table "properties_tags", id: false, force: :cascade do |t|
+      t.integer  "property_id"
+      t.integer  "tag_id"
+    end
+    
+    create_table "tags", force: :cascade do |t|
+      t.string   "name",                 limit: 255
+    end
   end
-
-  test '::sort(:has_many_relationship => {:column => :desc})' do
-    query = Property.sort(:addresses => {:id => :desc})
-
-    assert_equal(<<-SQL.strip.gsub(/\s+/, ' '), query.to_sql.gsub(/\s+/, ' '))
-      SELECT "properties".*, "addresses"."id" FROM "properties"
-      INNER JOIN "addresses" ON "addresses"."property_id" = "properties"."id"
-      ORDER BY "addresses"."id" DESC
-    SQL
+  
+  class Tag < ActiveRecord::Base
+    has_and_belongs_to_many :properties
   end
-
-  test '::sort(:has_many_relationship => {:column => {:asc => :nulls_first}})' do
-    query = Property.sort(:addresses => {:id => {:asc => :nulls_first}})
-
-    assert_equal(<<-SQL.strip.gsub(/\s+/, ' '), query.to_sql.gsub(/\s+/, ' '))
-      SELECT "properties".*, "addresses"."id" FROM "properties"
-      INNER JOIN "addresses" ON "addresses"."property_id" = "properties"."id"
-      ORDER BY "addresses"."id" ASC NULLS FIRST
-    SQL
+  
+  class Property < ActiveRecord::Base
+    has_and_belongs_to_many :tags
   end
-
-  test '::sort(:has_many_relationship => {:column => {:desc => :nulls_last}})' do
-    query = Property.sort(:addresses => {:id => {:desc => :nulls_last}})
-
-    assert_equal(<<-SQL.strip.gsub(/\s+/, ' '), query.to_sql.gsub(/\s+/, ' '))
-      SELECT "properties".*, "addresses"."id" FROM "properties"
-      INNER JOIN "addresses" ON "addresses"."property_id" = "properties"."id"
-      ORDER BY "addresses"."id" DESC NULLS LAST
-    SQL
-  end
-
+  
   test '::sort(:habtm_relationship => :column)' do
     query = Property.sort(:tags => :name)
 
@@ -101,6 +84,7 @@ class SortRelationTest < ActiveSupport::TestCase
     Property.create!(name: 'single', tags: [Tag.create!(name: 'Mall')])
 
     records = Property.includes(:tags).sort(:tags => {:name => :asc}).to_a
+
     names = records.map(&:name)
     assert_equal names.uniq, names
     assert_equal ['multi', 'single'], names
@@ -108,34 +92,5 @@ class SortRelationTest < ActiveSupport::TestCase
     assert records.all? { |record| record.association(:tags).loaded? }
     assert_equal ['Apex', 'Zoo'], records.first.tags.map(&:name).sort
   end
-
-  test '::sort(:belongs_to_relationship => {:column => :desc})' do
-    query = Address.sort(:property => {:id => :desc})
-
-    assert_equal(<<-SQL.strip.gsub(/\s+/, ' '), query.to_sql.gsub(/\s+/, ' '))
-      SELECT "addresses".*, "properties"."id" FROM "addresses"
-      LEFT OUTER JOIN "properties" ON "properties"."id" = "addresses"."property_id"
-      ORDER BY "properties"."id" DESC
-    SQL
-  end
-  
-  test '::sort(:belongs_to_relationship => {:column => :desc}).distinct_on' do
-    query = Address.sort(:property => {:id => :desc}).distinct_on(:id)
-
-    assert_equal(<<-SQL.strip.gsub(/\s+/, ' '), query.to_sql.gsub(/\s+/, ' '))
-      SELECT DISTINCT ON ( "addresses"."id" ) "addresses".*, "properties"."id" FROM "addresses"
-      LEFT OUTER JOIN "properties" ON "properties"."id" = "addresses"."property_id"
-      ORDER BY "properties"."id" DESC
-    SQL
-  end
-
-  # test 'SunstoneRecord::sort(:belongs_to_relationship => {:column => :desc})' do
-  #   webmock(:get, "/points", limit: 1, order: [{lines: {id: :desc}}]).to_return({
-  #     body: [{id: 42}].to_json
-  #   })
-
-  #   assert_equal 42, Point.sort(:line => {:id => :desc}).first.id
-  # end
-
 
 end
